Alternative implementation of chained completion
The original implementation does not work for me (Vim 7.3 on Win7 x64 / Ubuntu 12.04). Fallback completion menu is not shown until I enter another character. Not sure whether it is related to other plugins

I think this implementation is more reliable, because fallback completion won't be delayed and no need to mess with updatetime.