fix longest enhancement to not suppress "Pattern not found" message
with the longest enhancement enabled, the keyword completion "Pattern
not found" message was suppressed, seemingly due to CursorMovedI firing
even though the cursor hadn't moved. So to work around the issue, save
the col where the longest enhancement capturing starts and don't force
exiting of completion mode if the col hasn't changed when the
CursorMovedI fires.

addresses #38