track when supertab initiates the current completion mode
For the case where keyword completion is initiated without supertab's
knowledge, using <c-n>, then ideally <tab> should advance down the list,
not up. To avoid re-adding <c-n>/<c-p> mappings, supertab should set a
flag when it initiates the completio mode, and handle deciding whether
<c-n>/<c-p> move down or up the list accordingly. Note that this
solution does not solve the case where the completion was manually
started with <c-p> since supertab will advance down the list in this
case, but this may be a better default than the advancing up on <c-n>
completions.

fixes #113